refactor(cypress-integration): Renamed folder to auth-react
